KMeansClustering

Documentation for KMeansClustering.

KMeansClustering.CentroidCalculator

KMeansClustering.ClusterInit

KMeansClustering.EuclideanMeanCentroid

KMeansClustering.EuclideanNorm

KMeansClustering.KMeansPPInit

KMeansClustering.Norm

KMeansClustering.UniformRandomInit

KMeansClustering.KMeans

KMeansClustering.buildClusters

KMeansClustering.CentroidCalculator — Method

(c::CentroidCalculator{V})(samples::AbstractVector{V}, norm::Norm{V})::V where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Calculates the center from the samples using norm. Overwrite in your subtypes of CentroidCalculator

KMeansClustering.ClusterInit — Method

(c::ClusterInit{V})(samples::AbstractVector{V}, k::Int64, norm::Norm{V})::Vector{V} where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Initializes k cluster centers from samples using the cluster initialization method c. Overwrite in your subtypes of ClusterInit

KMeansClustering.EuclideanMeanCentroid — Method

(c::EuclideanMeanCentroid{V})(samples::AbstractVector{V}, norm::Norm{V})::V where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Calculates the center from the samples using the standard euclidean mean.

KMeansClustering.EuclideanNorm — Method

(c::EuclideanNorm{V})(x::V)::T where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Calculates the euclidean norm of x.

KMeansClustering.KMeansPPInit — Method

(c::KMeansPPInit{V})(samples::AbstractVector{V}, k::Int64, norm::Norm{V})::Vector{V} where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Perform K-means++ initialization to select initial cluster centers.

KMeansClustering.Norm — Method

(c::Norm{V})(x::V)::T where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Calculates the norm of x. Overwrite in your own subtypes of Norm.

KMeansClustering.UniformRandomInit — Type

(c::UniformRandomInit{V})(samples::AbstractVector{V}, k::Int64, norm::Norm{V})::Vector{V} where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Initializes k cluster centers from samples using a uniform random distribution, over the bounding hyperrectangle of the samples

KMeansClustering.KMeans — Method

KMeans(x::AbstractVector{V}, k::Int64; init::ClusterInit{V}=UniformRandomInit{V}(), max_iter=300, tol=0.0001, algorithm::KMeansAlgorithm=Lloyd)::Dict{V, Vector{V}}

Perform K-means clustering on the data x with k clusters.

Arguments:

x: Input data as an abstract vector of type V.

k: Number of clusters.

init: Cluster initialization method. Default is UniformRandomInit.

max_iter: Maximum number of iterations. Default is 300.

tol: Tolerance for convergence. Default is 0.0001.

algorithm: K-means algorithm to use. Default is Lloyd.

centroid: Used to calculate center of each cluster. Default EuclideanMeanCentroid

norm: Used to assign clusters to samples. Default EuclideanNorm

Returns a dictionary mapping each cluster center to its assigned samples.

KMeansClustering.buildClusters — Method

buildClusters(xs::AbstractVector{V}, init::AbstractVector{V}, norm::Norm{V})::Vector{Vector{V}} where {T<:NonInteger,N,V<:Union{T, AbstractArray{T,N}}}

Assigns each sample in xs to the nearest cluster center in init.

Returns a vector of clusters, where each cluster is a vector of samples.